국가 상세 페이지: 동적 라우팅에서 정적 생성으로 전환
- learn-react-and-next/section5/pages/country/[code].js 파일을 수정하여 국가 상세 페이지를 구현하는 방식을 변경했습니다.
- 기존에는 getServerSideProps 를 사용하여 요청마다 서버에서 데이터를 fetch 했던 방식에서
- 이제는 getStaticPaths 와 getStaticProps 를 사용하여 정적 생성 방식으로 변경했습니다.
- 이 변경을 통해 페이지 로딩 성능 향상과 SEO 개선 효과를 기대할 수 있습니다.
